Multi-chain support part 2 (#66)
Provides the ability for servers to support multiple different payment options, and for the clients to handle those and choose an available one.

examples/multichain has an example server that accepts payment via an ATXP Account with support for both Base and Solana, and a client that makes payments with both Solana and Base EOA wallets.