Batch payment example, implementation, and tests (#75)
Adds support for batch payments - payments that are more than the price of a single tool call. This allows the server to effectively request pre-payment for multiple tool calls in advance.

This is implemented via a new minimumPayment field on atxpExpress, which is the smallest payment that will be asked of callers. If not provided, the caller will always be asked to pay exactly the amount in requirePayment. If minimumPayment is larger, the user will be asked to pay that instead. To prevent footguns, this is limited to $1 at a maximum.

A new /examples/batch-payment example is included.